✨ 기능: payments 결제 관련 리팩토링, 에러 코드 작성, amount type int -> Long 변경

+++ backend/src/main/java/com/backend/payment/dto/PaymentResponse.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record PaymentResponse(
         String orderId,
         String paymentKey,
-        int amount,
+        Long amount,
         PaymentStatus status,
         LocalDateTime paidAt
 ) {
