+++ backend/src/main/java/com/backend/payment/dto/TossPaymentResponse.java
@@ -4,12 +4,12 @@ import java.time.OffsetDateTime;
 
 public record TossPaymentResponse(
         String paymentKey,
-        int totalAmount,
+        Long totalAmount,
         String method,
         OffsetDateTime approvedAt
 
 ) {
-    public static TossPaymentResponse success(String paymentKey, int amount, String method,OffsetDateTime approvedAt){
+    public static TossPaymentResponse success(String paymentKey, Long amount, String method,OffsetDateTime approvedAt){
         return new TossPaymentResponse(
                 paymentKey,
                 amount,
